chore(web): upgrade to Next 16.2.9 + OpenNext 1.20.1, apply Next best practices
Upgrade to the latest STABLE Next.js (16.3 is preview-only and its
cacheComponents/partialPrefetching features are not OpenNext/Cloudflare-safe),
and adopt instant-navigation + best-practice patterns achievable on stable.

Versions:
- next 16.2.6 -> 16.2.9; @next/{mdx,third-parties,bundle-analyzer} -> 16.2.9
- @opennextjs/cloudflare 1.19.6 -> 1.20.1; wrangler ^4.68 -> ^4.86 (4.105)

Client/server boundaries (remove unnecessary 'use client', fix CSR bailout):
- /todos and /auth/link-platform read useSearchParams in a client page, which
  bails the whole subtree out of static rendering. Convert both to server
  shells that await searchParams/params and pass props to a client child.
- Convert todos index/completed/label/priority/project pages to Server
  Components (they only built filters from params).

Correctness:
- use-cases/[slug]: redirect() was called inside try/catch, so its NEXT_REDIRECT
  control-flow throw was swallowed (or converted to notFound). Return the
  redirect target and call redirect() after the try; extract the lookup into a
  helper to keep complexity in check.

Instant navigation (Cloudflare/OpenNext streams these natively):
- Add loading.tsx streaming fallbacks for (landing) and (main) route groups.
- Add error.tsx segment boundaries (shared RouteError) so a render error is
  contained to its route group instead of replacing the whole app shell.
- marketplace/[slug]: run the two independent fetches with Promise.all.

Other: replace a raw <img> with next/image; log (not swallow) the optional
link-token-info fetch failure.

Note: middleware.ts and the OG routes' edge runtime are intentionally kept
(OpenNext-Cloudflare requires edge middleware; proxy.ts forces Node).

Entire-Checkpoint: 5a87ba7c75c6
