+++ apps/web/package.json
@@ -78,10 +78,10 @@
     "@mdx-js/react": "^3.1.0",
     "@microsoft/fetch-event-source": "^2.0.1",
     "@modelcontextprotocol/sdk": "^1.27.1",
-    "@next/mdx": "^16.2.6",
-    "@next/third-parties": "^16.2.6",
+    "@next/mdx": "^16.2.9",
+    "@next/third-parties": "^16.2.9",
     "@number-flow/react": "^0.6.0",
-    "@opennextjs/cloudflare": "^1.19.6",
+    "@opennextjs/cloudflare": "^1.20.1",
     "@openuidev/react-headless": "^0.8.2",
     "@openuidev/react-lang": "^0.2.6",
     "@openuidev/react-ui": "^0.11.9",
@@ -176,7 +176,7 @@
     "mermaid": "^11.10.1",
     "moment-timezone": "^0.6.0",
     "motion": "^12.23.12",
-    "next": "16.2.6",
+    "next": "16.2.9",
     "next-intl": "^4.8.3",
     "next-llms-txt": "^1.0.2",
     "onnxruntime-web": "^1.20.0",
@@ -224,7 +224,7 @@
   },
   "devDependencies": {
     "@biomejs/biome": "^2.3.7",
-    "@next/bundle-analyzer": "^16.2.6",
+    "@next/bundle-analyzer": "^16.2.9",
     "@types/dagre": "^0.7.53",
     "@types/node": "^24.3.0",
     "@types/react": "^19.2.14",
@@ -234,7 +234,7 @@
     "depcheck": "^1.4.7",
     "schema-dts": "^1.1.5",
     "typescript": "^5.9.3",
-    "wrangler": "^4.68.0"
+    "wrangler": "^4.86.0"
   },
   "packageManager": "pnpm@10.18.3"
 }
