+++ apps/web/src/app/[locale]/(main)/error.tsx
@@ -1,0 +1,3 @@
+"use client";
+
+export { default } from "@/components/shared/RouteError";
