+++ apps/web/src/app/[locale]/(main)/loading.tsx
@@ -1,0 +1,5 @@
+import SuspenseLoader from "@/components/shared/SuspenseLoader";
+
+export default function Loading() {
+  return <SuspenseLoader />;
+}
