+++ apps/web/src/app/[locale]/(main)/todos/completed/page.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import TodoListPage from "@/features/todo/components/TodoListPage";
 
 export default function CompletedTodosPage() {
