+++ apps/web/src/app/[locale]/(main)/todos/page.tsx
@@ -1,38 +1,28 @@
-"use client";
-
-import { useSearchParams } from "next/navigation";
-import { useMemo } from "react";
-
 import TodoListPage from "@/features/todo/components/TodoListPage";
 import { Priority, type TodoFilters } from "@/types/features/todoTypes";
 
-export default function TodosPage() {
-  const searchParams = useSearchParams();
+interface TodosPageProps {
+  searchParams: Promise<{
+    project?: string;
+    priority?: string;
+    completed?: string;
+  }>;
+}
 
-  const projectId = searchParams.get("project");
-  const priority = searchParams.get("priority");
-  const completedParam = searchParams.get("completed");
+export default async function TodosPage({ searchParams }: TodosPageProps) {
+  const { project, priority, completed } = await searchParams;
 
-  const filters = useMemo((): TodoFilters => {
-    const urlFilters: TodoFilters = {};
+  const filters: TodoFilters = {};
 
-    if (projectId) {
-      urlFilters.project_id = projectId;
-    }
+  if (project) {
+    filters.project_id = project;
+  }
 
-    if (priority) {
-      const priorityValue = Object.values(Priority).includes(
-        priority as Priority,
-      )
-        ? (priority as Priority)
-        : undefined;
-      if (priorityValue) urlFilters.priority = priorityValue;
-    }
+  if (priority && Object.values(Priority).includes(priority as Priority)) {
+    filters.priority = priority as Priority;
+  }
 
-    urlFilters.completed = completedParam === "true";
-
-    return urlFilters;
-  }, [projectId, priority, completedParam]);
+  filters.completed = completed === "true";
 
   return <TodoListPage filters={filters} />;
 }
