+++ apps/web/src/app/[locale]/(main)/todos/priority/[priority]/page.tsx
@@ -1,14 +1,18 @@
-"use client";
-
-import { useParams } from "next/navigation";
-import { useMemo } from "react";
-
 import TodoListPage from "@/features/todo/components/TodoListPage";
 import type { Priority } from "@/types/features/todoTypes";
 
-export default function PriorityTodosPage() {
-  const params = useParams();
-  const priority = params.priority as Priority;
-  const filters = useMemo(() => ({ priority, completed: false }), [priority]);
-  return <TodoListPage filters={filters} />;
+interface PriorityTodosPageProps {
+  params: Promise<{ priority: string }>;
+}
+
+export default async function PriorityTodosPage({
+  params,
+}: PriorityTodosPageProps) {
+  const { priority } = await params;
+
+  return (
+    <TodoListPage
+      filters={{ priority: priority as Priority, completed: false }}
+    />
+  );
 }
