+++ package.json
@@ -48,7 +48,7 @@
     "jscpd": "^4.2.2",
     "knip": "^5",
     "madge": "^8.0.0",
-    "next": "16.2.6",
+    "next": "16.2.9",
     "npm-package-json-lint": "^10.4.0",
     "nx": "^22.2.3",
     "publint": "^0.3.21",
